Macro Communication are working to specifications

+++ ControlBoxDriver/communications/communications.hpp
@@ -31,8 +31,8 @@ public:
 	void handle_emergency_stop();
 	void send_stop_macro ();
 	void ping_robot();
-	int8_t get_requested_macro ();
-	void handle_macro_request (uint8_t);
+	int16_t get_requested_macro ();
+	void handle_macro_request (uint16_t);
 	void handle_manual_command ();
 	void stop_all_motors ();
 	void send (const Message [], uint8_t);
